+++ simulator/risc_v/riscv_register/t/unit_test.cpp
@@ -19,7 +19,7 @@ static_assert(RISCVRegister::MAX_REG == 32);
 // Testing methods of the class
 TEST_CASE( "RISCV_registers: Size_t_converters")
 {
-    for ( size_t i = 0; i < 32; ++i)
+    for ( uint8 i = 0; i < 32; ++i)
     {
         CHECK( RISCVRegister::from_cpu_index( i).to_rf_index() == i);
     }
@@ -27,7 +27,7 @@ TEST_CASE( "RISCV_registers: Size_t_converters")
 
 TEST_CASE( "RISCV_registers: Equal")
 {
-    for ( size_t i = 0; i < 32; ++i)
+    for ( uint8 i = 0; i < 32; ++i)
     {
         CHECK( RISCVRegister::from_cpu_index( i) == RISCVRegister::from_cpu_index( i));
         if (i > 0) {
@@ -40,7 +40,7 @@ TEST_CASE( "RISCV_registers: no_mips")
 {
     auto reg_hi = RISCVRegister::mips_hi();
     auto reg_lo = RISCVRegister::mips_lo();
-    for( size_t i = 0; i < 32; ++i)
+    for( uint8 i = 0; i < 32; ++i)
     {
         // Ensure that there are no mips regs
         CHECK( RISCVRegister::from_cpu_index( i).to_rf_index() != reg_hi.to_rf_index());
